requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
langdetect>=1.0.9
deep-translator>=1.11.4
SpeechRecognition>=3.10.4
faster-whisper>=1.2.1
